# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProjectTnW - Multi-tenant School Transportation SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/current:
    get:
      operationId: getCurrentTenant
      tags: [tenants]
      summary: Get current tenant details
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

    patch:
      operationId: updateCurrentTenant
      tags: [tenants]
      summary: Update current tenant settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # ─── USERS (within tenant) ───
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role in the tenant
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"

    delete:
      operationId: removeUserFromTenant
      tags: [users]
      summary: Remove a user from the tenant
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ─── INVITATIONS ───
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List pending invitations
      responses:
        "200":
          description: Invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"

    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Send an invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /invitations/{invitationId}:
    delete:
      operationId: deleteInvitation
      tags: [invitations]
      summary: Cancel an invitation
      parameters:
        - in: path
          name: invitationId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationBody"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── VEHICLES ───
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        tenant:
          $ref: "#/components/schemas/Tenant"
        role:
          type: string
      required:
        - token
        - user
        - tenant
        - role
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        isActive:
          type: boolean
        isSuperAdmin:
          type: boolean
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - email
        - isActive
        - isSuperAdmin
...
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        isDemo:
          type: ["boolean", "null"]
        plan:
          type: ["string", "null"]
        billingStatus:
          type: ["string", "null"]
        trialEndsAt:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - slug
    UpdateTenantBody:
      type: object
      properties:
        name:
          type: string
        settings:
          type: object

    # ─── USERS ───
    TenantUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: ["boolean", "null"]
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - email
        - role
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
      required:
        - role

    # ─── INVITATIONS ───
    Invitation:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        expiresAt:
          type: string
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - email
        - role
        - expiresAt
    CreateInvitationBody:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
      required:
        - email
        - role
    AcceptInvitationBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - token

    # ─── VEHICLES ───
...